backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
mysql-connector-python==9.0.0
reportlab==4.2.2
